Move the routines Init_ExtAcc/PotAuxArray* into CPU_ExtAcc_*.cpp
