+++ src/TestProblem/Hydro/Plummer/ExtAcc_Plummer.cpp
@@ -84,4 +84,38 @@ void SetCPUExtAcc_Plummer( ExtAcc_t &CPUExtAcc_Ptr )
 
 
 
+#ifndef __CUDACC__
+
+extern double Plummer_Rho0;
+extern double Plummer_R0;
+extern double Plummer_Center[3];
+extern double Plummer_ExtMFrac;
+
+//-------------------------------------------------------------------------------------------------------
+// Function    :  Init_ExtAccAuxArray_Plummer
+// Description :  Set the auxiliary array ExtAcc_AuxArray[] used by ExtAcc_Plummer()
+//
+// Note        :  1. AuxArray[] has the size of EXT_ACC_NAUX_MAX defined in Macro.h (default = 10)
+//
+// Parameter   :  AuxArray : Array to be filled up
+//
+// Return      :  AuxArray[]
+//-------------------------------------------------------------------------------------------------------
+void Init_ExtAccAuxArray_Plummer( double AuxArray[] )
+{
+
+// acceleration = -G*Mtot*r/(r^2+R0^2)^(3/2)
+   const double Mtot = (4.0/3.0)*M_PI*CUBE(Plummer_R0)*Plummer_Rho0*Plummer_ExtMFrac;
+
+   AuxArray[0] = Plummer_Center[0];    // x coordinate of the external acceleration center
+   AuxArray[1] = Plummer_Center[1];    // y ...
+   AuxArray[2] = Plummer_Center[2];    // z ...
+   AuxArray[3] = SQR( Plummer_R0 );    // scale_radius^2
+   AuxArray[4] = -NEWTON_G*Mtot;       // -G*total_mass
+
+} // FUNCTION : Init_ExtAccAuxArray_Plummer
+#endif // #ifndef __CUDACC__
+
+
+
 #endif // #ifdef GRAVITY
